+++ Logic_ex.c
@@ -1,7 +1,7 @@
 /*!\file Logic_ex.c
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2022, SMFSW
-** \brief Extensions for logic variable handling
+** \brief Extensions for logic variables (R/W)
 **/
 /****************************************************************/
 #include "Logic_ex.h"